// src/features/editableProfileCard/model/consts/consts.ts
export enum ValidateProfileError {
  NO_DATA_USER_FIRSTNAME = 'EPMTY_USER_FIRSTNAME',
  NO_DATA_USER_LASTNAME = 'EPMTY_USER_LASTNAME',
  INCORRECT_AGE = 'INCORRECT_AGE',
  INCORRECT_COUNTRY = 'INCORRECT_COUNTRY',
  INCORRECT_CITY = 'INCORRECT_CITY',
  INCORRECT_CURRENCY = 'INCORRECT_CURRENCY',
  SERVER_ERROR = 'SERVER_ERROR',
  LONG_USERNAME = 'LONG_USERNAME',
  NO_DATA = 'NO_DATA',
}